+++ src/settings/OrderTemplates/OrderTemplatesEditor/OrderTemplatesEditor.test.js
@@ -20,6 +20,18 @@ jest.mock('@folio/stripes/components', () => ({
   expandAllSections: jest.fn(),
   Layer: jest.fn(({ children }) => <>{children}</>),
 }));
+jest.mock('@folio/stripes/smart-components', () => ({
+  ...jest.requireActual('@folio/stripes/smart-components'),
+  EditCustomFieldsRecord: jest.fn().mockImplementation(({ accordionId }) => {
+    if (accordionId === 'poCustomFields') {
+      return <div>EditCustomFieldsRecord_PO</div>;
+    } else if (accordionId === 'polCustomFields') {
+      return <div>EditCustomFieldsRecord_POL</div>;
+    }
+
+    return <div>EditCustomFieldsRecord</div>;
+  }),
+}));
 
 const defaultProps = {
   close: jest.fn(),
@@ -110,6 +122,8 @@ describe('OrderTemplatesEditor', () => {
       expect(screen.getByText('ui-orders.settings.orderTemplates.accordion.polFundDistribution')).toBeInTheDocument();
       expect(screen.getByText('ui-orders.settings.orderTemplates.accordion.polLocation')).toBeInTheDocument();
       expect(screen.getByText('ui-orders.settings.orderTemplates.accordion.polTags')).toBeInTheDocument();
+      expect(screen.getByText('EditCustomFieldsRecord_PO')).toBeInTheDocument();
+      expect(screen.getByText('EditCustomFieldsRecord_POL')).toBeInTheDocument();
     });
   });